finished

+++ libft/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: hyejeong <hyejeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/23 12:45:18 by hyejeong          #+#    #+#             */
-/*   Updated: 2022/11/23 12:54:14 by hyejeong         ###   ########.fr       */
+/*   Updated: 2022/11/29 23:34:48 by hyejeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
