+++ libft/ft_strchr.c
@@ -6,14 +6,16 @@
 /*   By: hyejeong <hyejeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 11:39:17 by hyejeong          #+#    #+#             */
-/*   Updated: 2022/11/28 15:35:08 by hyejeong         ###   ########.fr       */
+/*   Updated: 2022/11/29 23:35:26 by hyejeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 char	*ft_strchr(const char *s, int c)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	str = (char *)s;
 	i = 0;
@@ -28,5 +30,3 @@ char	*ft_strchr(const char *s, int c)
 	else
 		return (0);
 }
-//문자열 내에 일치하는 문자가 있는지 검사하는 함수
-//처음 만나는 지점의 포인터 리턴
