+++ libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: hyejeong <hyejeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 11:39:40 by hyejeong          #+#    #+#             */
-/*   Updated: 2022/11/23 13:39:49 by hyejeong         ###   ########.fr       */
+/*   Updated: 2022/11/29 23:36:07 by hyejeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strdup(const char *s)
 {
-	int		i;
-	int		s_len;
+	size_t	i;
+	size_t	s_len;
 	char	*str;
 
 	i = 0;
