+++ libft/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: hyejeong <hyejeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 12:52:16 by hyejeong          #+#    #+#             */
-/*   Updated: 2022/11/28 15:35:25 by hyejeong         ###   ########.fr       */
+/*   Updated: 2022/11/29 23:37:50 by hyejeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	char	*str;
-	int		s_len;
-	int		i;
+	size_t	s_len;
+	size_t	i;
 
 	s_len = ft_strlen(s);
 	str = (char *)malloc(sizeof(char) * (s_len + 1));
