+++ libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: hyejeong <hyejeong@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/23 00:15:02 by hyejeong          #+#    #+#             */
-/*   Updated: 2022/11/24 18:03:22 by hyejeong         ###   ########.fr       */
+/*   Updated: 2022/11/29 23:38:22 by hyejeong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static int	filtering(char c, char const *set)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (set[i] != '\0')
